tests/wpt/web-platform-tests/html/semantics/popovers/popover-animation-corner-cases.tentative.html


<!DOCTYPE html>
<meta charset="utf-8">
<link rel=author href="mailto:masonf@chromium.org">
<link rel=help href="https://open-ui.org/components/popup.research.explainer">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/popover-utils.js"></script>

<body>
<style>
.animation { opacity: 0; }
.animation:open { opacity: 1; }
.animation:not(:open) { animation: fade-out 1000s; }
@keyframes fade-out {
  from { opacity: 1; }
  to { opacity: 0; }
}

.animation>div>div { left: 0; }
.animation:not(:open)>div>div { animation: rotate 1000s; color:red;}
@keyframes rotate {
  from { transform: rotate(0); }
  to { transform: rotate(360deg); }
}

[popover] { top: 200px; }
[popover]::backdrop { background-color: rgba(255,0,0,0.2); }
</style>

<script>
function createPopover(t,type) {
  const popover = document.createElement('div');
  popover.popover = 'auto';
  popover.classList = type;
  const div = document.createElement('div');
  const descendent = div.appendChild(document.createElement('div'));
  descendent.appendChild(document.createTextNode("Descendent element"));
  popover.append("This is a pop up",div);
  document.body.appendChild(popover);
  t.add_cleanup(() => popover.remove());
  return {popover, descendent};
}
promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'animation');
  assert_false(isElementVisible(popover));
  assert_equals(descendent.parentElement.parentElement,popover);
  assert_true(popover.matches(':closed'));
  assert_false(popover.matches(':open'));
  popover.showPopover();
  assert_false(popover.matches(':closed'));
  assert_true(popover.matches(':open'));
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  popover.hidePopover();
  const animations = popover.getAnimations({subtree: true});
  assert_equals(animations.length,2,'There should be two animations running');
  assert_false(popover.matches(':open'),'popover should not match :open as soon as hidden');
  assert_false(popover.matches(':closed'),'popover should not match :closed until animations complete');
  assert_true(isElementVisible(popover),'but animations should keep the popover visible');
  assert_true(isElementVisible(descendent),'The descendent should also be visible');
  await waitForRender();
  await waitForRender();
  assert_equals(popover.getAnimations({subtree: true}).length,2,'The animations should still be running');
  assert_true(isElementVisible(popover),'Popover should still be visible due to animation');
  animations.forEach(animation => animation.finish()); // Force the animations to finish
  await waitForRender(); // Wait one frame
  assert_true(popover.matches(':closed'),'The pop up should now match :closed');
  assert_false(popover.matches(':open'),'The pop up still shouldn\'t match :open');
  assert_false(isElementVisible(popover),'The pop up should now be invisible');
  assert_false(isElementVisible(descendent),'The descendent should also be invisible');
  assert_equals(popover.getAnimations({subtree: true}).length,0);
},'Descendent animations should keep the pop up visible until the animation ends');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  assert_equals(popover.classList.length, 0);
  assert_false(isElementVisible(popover));
  popover.showPopover();
  assert_true(popover.matches(':open'));
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  // Start an animation on the popover and its descendent.
  popover.animate([{opacity: 1},{opacity: 0}],{duration: 1000000,iterations: 1});
  descendent.animate([{transform: 'rotate(0)'},{transform: 'rotate(360deg)'}],1000000);
  assert_equals(popover.getAnimations({subtree: true}).length,2);
  // Then hide the popover.
  popover.hidePopover();
  assert_false(popover.matches(':open'),'pop up should not match :open as soon as hidden');
  assert_true(popover.matches(':closed'),'pop up should match :closed immediately');
  assert_equals(popover.getAnimations({subtree: true}).length,2,'animations should still be running');
  await waitForRender();
  assert_equals(popover.getAnimations({subtree: true}).length,2,'animations should still be running');
  assert_false(isElementVisible(popover),'Pre-existing animations should not keep the pop up visible');
},'Pre-existing animations should *not* keep the pop up visible until the animation ends');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  popover.showPopover();
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  let animation;
  popover.addEventListener('beforetoggle', (e) => {
    if (e.newState !== "closed")
      return;
    animation = popover.animate([{opacity: 1},{opacity: 0}],1000000);
  });
  assert_equals(popover.getAnimations({subtree: true}).length,0,'There should be no animations yet');
  popover.hidePopover();
  assert_equals(popover.getAnimations({subtree: true}).length,1,'the hide animation should now be running');
  assert_true(!!animation);
  assert_true(isElementVisible(popover),'The animation should keep the popover visible');
  animation.finish();
  await waitForRender();
  assert_false(isElementVisible(popover),'Once the animation ends, the popover is hidden');
},'It should be possible to use the "beforetoggle" event handler to animate the hide');


promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  const dialog = document.body.appendChild(document.createElement('dialog'));
  t.add_cleanup(() => dialog.remove());
  popover.showPopover();
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  let animation;
  popover.addEventListener('beforetoggle', (e) => {
    if (e.newState !== "closed")
      return;
    animation = popover.animate([{opacity: 1},{opacity: 0}],1000000);
  });
  assert_equals(popover.getAnimations({subtree: true}).length,0,'There should be no animations yet');
  dialog.showModal(); // Force hide the popover
  await waitForRender();
  assert_equals(popover.getAnimations({subtree: true}).length,1,'the hide animation should now be running');
  assert_true(isElementVisible(popover),'And the animation should keep the popover visible');
  animation.finish();
  await waitForRender();
  assert_false(isElementVisible(popover),'Once the animation ends, the popover is hidden');
},'It should be possible to use the "beforetoggle" event handler to animate the hide, even when the hide is due to dialog.showModal');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  popover.showPopover();
  assert_true(isElementVisible(popover));
  popover.addEventListener('beforetoggle', (e) => e.preventDefault());
  popover.hidePopover();
  await waitForRender();
  assert_false(isElementVisible(popover),'Even if hide event is cancelled, the popover still closes');
},'toggle event cannot be cancelled');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'animation');
  assert_false(isElementVisible(popover));
  popover.showPopover();
  assert_false(popover.matches(':closed'));
  assert_true(popover.matches(':open'));
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  popover.popover = 'manual';
  const animations = popover.getAnimations({subtree: true});
  assert_equals(animations.length,2,'There should be two animations running');
  assert_false(popover.matches(':open'),'popover should not match :open as soon as hidden');
  assert_false(popover.matches(':closed'),'popover should not match :closed until animations complete');
  assert_true(isElementVisible(popover),'but animations should keep the popover visible');
  animations.forEach(animation => animation.finish()); // Force the animations to finish
  await waitForRender(); // Wait one frame
  assert_true(popover.matches(':closed'),'The pop up should now match :closed');
  assert_false(popover.matches(':open'),'The pop up still shouldn\'t match :open');
  assert_false(isElementVisible(popover),'The pop up should now be invisible');
},'Closing animations are triggered by changing the popover type');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  popover.showPopover();
  assert_true(isElementVisible(popover));
  assert_equals(popover.getAnimations({subtree: true}).length,0);
  popover.addEventListener('beforetoggle', (e) => {
    if (e.newState !== "closed")
      return;
    popover.animate([{opacity: 1},{opacity: 0}],1000000);
  });
  assert_equals(popover.getAnimations({subtree: true}).length,0,'There should be no animations yet');
  popover.hidePopover();
  await waitForRender();
  assert_equals(popover.getAnimations({subtree: true}).length,1,'the hide animation should now be running');
  assert_true(isElementVisible(popover),'The popover should still be visible because the animation hasn\'t ended.');
  const animation = popover.getAnimations({subtree: true})[0];

  animation.dispatchEvent(new Event('finish'));
  await waitForRender();
  assert_true(isElementVisible(popover),'Synthetic finish events should not stop the animation, so the popover should still be visible.');
  assert_equals(popover.getAnimations({subtree: true}).length,1,'the hide animation should still be running');

  animation.dispatchEvent(new Event('cancel'));
  await waitForRender();
  assert_true(isElementVisible(popover),'Synthetic cancel events should not stop the animation, so the popover should still be visible.');
  assert_equals(popover.getAnimations({subtree: true}).length,1,'the hide animation should still be running');
},'animation finish/cancel events must be trusted in order to finish closing the popover.');

promise_test(async (t) => {
  const {popover, descendent} = createPopover(t,'');
  popover.showPopover();
  popover.addEventListener('beforetoggle', (e) => {
    if (e.newState !== "closed")
      return;
    popover.animate([{opacity: 1},{opacity: 0}],1000000);
  });
  popover.hidePopover();
  await waitForRender();
  assert_true(isElementVisible(popover),'The popover should still be visible because the animation hasn\'t ended.');
  assert_equals(popover.getAnimations({subtree: true}).length,1,'There should be one animation running');
  const animation = popover.getAnimations({subtree: true})[0];
  let new_animation;
  const listener = () => {new_animation = popover.animate([{opacity: 1},{opacity: 0}],1000000)};
  animation.addEventListener('finish',listener,{capture:true});
  animation.addEventListener('cancel',listener,{capture:true});
  popover.addEventListener('animationfinish',listener,{capture:true});
  popover.addEventListener('animationcancel',listener,{capture:true});
  assert_true(isElementVisible(popover),'The popover should still be visible.');
  assert_equals(new_animation, undefined,'New animation should not be started yet.');
  animation.finish();
  await waitForRender(); // Wait one frame
  assert_true(popover.matches(':closed'),'The pop up should now match :closed');
  assert_false(popover.matches(':open'),'The pop up still shouldn\'t match :open');
  assert_false(isElementVisible(popover),'The pop up should now be invisible');
  assert_not_equals(new_animation, animation);
  assert_equals(popover.getAnimations({subtree: true})[0],new_animation,'The new animation should now be running');
},'Capturing event listeners can\'t affect popover animations.');
</script>
