LayoutTests/imported/w3c/web-platform-tests/html/semantics/popovers/popover-animation-corner-cases.tentative-expected.txt


FAIL Descendent animations should keep the pop up visible until the animation ends assert_false: expected false got true
FAIL Pre-existing animations should *not* keep the pop up visible until the animation ends assert_false: expected false got true
FAIL It should be possible to use the "beforetoggle" event handler to animate the hide assert_equals: the hide animation should now be running expected 1 but got 0
FAIL It should be possible to use the "beforetoggle" event handler to animate the hide, even when the hide is due to dialog.showModal assert_equals: the hide animation should now be running expected 1 but got 0
FAIL toggle event cannot be cancelled assert_false: Even if hide event is cancelled, the popover still closes expected false got true
FAIL Closing animations are triggered by changing the popover type assert_false: expected false got true
FAIL animation finish/cancel events must be trusted in order to finish closing the popover. assert_equals: the hide animation should now be running expected 1 but got 0
FAIL Capturing event listeners can't affect popover animations. assert_equals: There should be one animation running expected 1 but got 0
